requirements.txt
transformers==4.40.0
transformers[sentencepiece]
datasets
sacrebleu
rouge_score
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
torch
torchaudio
librosa==0.10.1
soundfile
pydub
numpy
scikit-learn
streamlit
openai
python-box==6.0.2
ensure==1.0.2
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
boto3
mypy-boto3-s3
